code/requirements.txt
anthropic>=0.25,<1.0
pandas>=2.0,<3.0
tqdm>=4.65,<5.0
scikit-learn>=1.3,<2.0
sentence-transformers>=2.2,<4.0
rank_bm25>=0.2,<1.0
numpy>=1.24,<3.0
python-dotenv>=1.0,<2.0
